Add user management page

+++ src/components/DeleteUserButton.tsx
@@ -1,18 +1,27 @@
 "use client";
 
-import { startTransition } from "react";
+import { useTransition } from "react";
 import { deleteUser } from "../server-functions/deleteUser";
+import { IconTrash } from "@tabler/icons-react";
+import { ActionIcon } from "@mantine/core";
 
 export function DeleteUserButton({ userId }: { userId: number }) {
+  const [pending, trackTransition] = useTransition();
+
   return (
-    <button
+    <ActionIcon
+      loading={pending}
+      variant="subtle"
+      color="red"
       onClick={() => {
-        startTransition(async () => {
-          await deleteUser(userId);
-        });
+        if (confirm("Opravdu smazat uživatele?")) {
+          trackTransition(async () => {
+            await deleteUser(userId);
+          });
+        }
       }}
     >
-      Smazat uživatele
-    </button>
+      <IconTrash />
+    </ActionIcon>
   );
 }
